Consolidate down a couple articles
Add the notes on Observability in control theory practice talk to the Bae article, and note how cool she is.
Add some notes to the bicycling readme. Not a whole lot though.
Finally consolidate the meta/setup.md and meta/duplicating.md articles.
Update the tooling page with validation tooling.